# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SEORx API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List all organizations for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create a new organization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationBody"
      responses:
        "201":
          description: Created organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationBody"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteOrganization
      tags: [organizations]
      summary: Delete organization
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{orgId}/members:
    get:
      operationId: listOrgMembers
      tags: [organizations]
      summary: List members of an organization
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrgMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: inviteOrgMember
      tags: [organizations]
      summary: Invite a member to an organization
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteOrgMemberBody"
      responses:
        "201":
          description: Invited member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgMember"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List generated reports
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: query
          schema:
            type: string
        - name: auditId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createReport
      tags: [reports]
      summary: Generate a report for an audit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete report
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /pagespeed/{auditId}:
    get:
      operationId: getPageSpeedResults
      tags: [pagespeed]
      summary: Get PageSpeed results for an audit
      security:
        - bearerAuth: []
      parameters:
        - name: auditId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PageSpeed results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageSpeedResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        plan:
          type: string
          enum: [free, starter, professional, enterprise]
        memberCount:
          type: integer
        clientCount:
          type: integer
        auditCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - plan
        - memberCount
        - clientCount
        - auditCount
        - createdAt
        - updatedAt
    CreateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
      required:
        - name
        - slug
    UpdateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
        plan:
          type: string
          enum: [free, starter, professional, enterprise]
    OrgMember:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        email:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, agency, client, viewer]
        joinedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - email
        - role
        - joinedAt
    InviteOrgMemberBody:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, agency, client, viewer]
      required:
        - email
        - role
...
    AuditIssue:
      type: object
      properties:
        id:
          type: string
        auditId:
          type: string
        url:
          type: string
        category:
          type: string
          enum: [meta, content, performance, links, structured_data, mobile, security, crawlability]
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        title:
          type: string
        description:
          type: string
        recommendation:
          type: string
          nullable: true
        aiRecommendation:
          type: string
          nullable: true
        priorityScore:
          type: number
        status:
          type: string
          enum: [open, approved, dismissed, fixed]
        approvedBy:
          type: string
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        affectedElement:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - auditId
        - url
        - category
        - severity
        - title
        - description
        - priorityScore
        - status
        - createdAt
...
    Report:
      type: object
      properties:
        id:
          type: string
        auditId:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        title:
          type: string
        format:
          type: string
          enum: [pdf, html, json]
        status:
          type: string
          enum: [generating, ready, failed]
        downloadUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - auditId
        - clientId
        - clientName
        - title
        - format
        - status
        - createdAt
        - updatedAt
    ReportDetail:
      allOf:
        - $ref: "#/components/schemas/Report"
        - type: object
          properties:
            summary:
              type: string
              nullable: true
            issueCount:
              type: integer
            topIssues:
              type: array
              items:
                $ref: "#/components/schemas/AuditIssue"
    CreateReportBody:
      type: object
      properties:
        auditId:
          type: string
        title:
          type: string
        format:
          type: string
          enum: [pdf, html, json]
          default: pdf
        includeAiSummary:
          type: boolean
          default: true
      required:
        - auditId
        - title
...
    PageSpeedResult:
      type: object
      properties:
        auditId:
          type: string
        url:
          type: string
        performanceScore:
          type: number
        accessibilityScore:
          type: number
        bestPracticesScore:
          type: number
        seoScore:
          type: number
        fcp:
          type: number
          description: First Contentful Paint (ms)
        lcp:
          type: number
          description: Largest Contentful Paint (ms)
        cls:
          type: number
          description: Cumulative Layout Shift
        tbt:
          type: number
          description: Total Blocking Time (ms)
        ttfb:
          type: number
          description: Time to First Byte (ms)
        fetchedAt:
          type: string
          format: date-time
      required:
        - auditId
        - url
        - performanceScore
        - accessibilityScore
        - bestPracticesScore
        - seoScore
        - fetchedAt